✨ Feat: MVC 상위 class를 ts로 변경(Core, Controller, Model, View)
