+++ src/core/Core.ts
@@ -1,6 +1,8 @@
 import Router from '../router';
 
 export default class Core {
+  $router: Router;
+
   constructor() {
     this.$router = Router.router;
   }
