+++ src/core/Model.ts
@@ -1,0 +1,5 @@
+import Core from './Core';
+
+export default class Model extends Core {
+  init() {}
+}
